introduce EditorActions as single behavior layer for editor inputs
